nothing

+++ app/src/main/java/com/xiaoyu/myweibo/bean/WeiBoDetailList.java
@@ -29,7 +29,7 @@ public class WeiBoDetailList {
     private int interval;
     private long max_id;
     private long next_cursor;
-    private int previous_cursor;
+    private long previous_cursor;
     private long since_id;
     private int total_number;
     private int uve_blank;
@@ -116,11 +116,11 @@ public class WeiBoDetailList {
         this.next_cursor = next_cursor;
     }
 
-    public int getPrevious_cursor() {
+    public long getPrevious_cursor() {
         return previous_cursor;
     }
 
-    public void setPrevious_cursor(int previous_cursor) {
+    public void setPrevious_cursor(long previous_cursor) {
         this.previous_cursor = previous_cursor;
     }
